\name{circos.genomicTrack}
\alias{circos.genomicTrack}
\title{
Create a track for genomic graphics
}
\description{
Create a track for genomic graphics
}
\usage{
circos.genomicTrack(...)
}
\arguments{

  \item{...}{Pass to \code{\link{circos.genomicTrackPlotRegion}}.}

}
\details{
shortcut function of \code{\link{circos.genomicTrackPlotRegion}}.
}
\examples{
# There is no example
NULL

}
