\name{circos.heatmap.get.x}
\alias{circos.heatmap.get.x}
\title{
Get the x-position for heatmap rows
}
\description{
Get the x-position for heatmap rows
}
\usage{
circos.heatmap.get.x(row_ind)
}
\arguments{

  \item{row_ind}{A vector of row indicies.}

}
\value{
A three-column data frame of the sector, the x-positions on the corresponding sectors, and the original row indicies.
}
\examples{
# There is no example
NULL

}
