\name{circos.info}
\alias{circos.info}
\title{
  Label the sector index and the track index of each cell  


}
\description{
  Label the sector index and the track index of each cell  


}
\usage{
circos.info()
}
\details{
  Draw the index of the sector and the track for each cell on the figure. This function can help you to find the coordinates of cells.  


}
