\name{circos.initializeCircularGenome}
\alias{circos.initializeCircularGenome}
\title{
Initialize a layout for circular genome
}
\description{
Initialize a layout for circular genome
}
\usage{
circos.initializeCircularGenome(name, genome_size, plotType = "axis", ...)
}
\arguments{

  \item{name}{Name of the genome (or the "chromosome name").}
  \item{genome_size}{Size of the genome}
  \item{plotType}{Pass to \code{\link{circos.genomicInitialize}}.}
  \item{...}{All goes to \code{\link{circos.genomicInitialize}}.}

}
\examples{
# There is no example
NULL

}
