\name{circos.track}
\alias{circos.track}
\title{
Create plotting regions for a whole track
}
\description{
Create plotting regions for a whole track
}
\usage{
circos.track(...)
}
\arguments{

  \item{...}{pass to \code{\link{circos.trackPlotRegion}}}

}
\details{
Shortcut function of \code{\link{circos.trackPlotRegion}}.
}
\examples{
# There is no example
NULL

}
