\name{circos.update}
\alias{circos.update}
\title{
Create plotting regions for a whole track
}
\description{
Create plotting regions for a whole track
}
\usage{
circos.update(...)
}
\arguments{

  \item{...}{pass to \code{\link{circos.updatePlotRegion}}}

}
\details{
shortcut function of \code{\link{circos.updatePlotRegion}}.
}
\examples{
# There is no example
NULL

}
