\name{convert_width}
\alias{convert_width}
\title{
Convert units
}
\description{
Convert units
}
\usage{
convert_width(x, unit = c("mm", "cm", "inches"))
}
\arguments{

  \item{x}{a numeric vector}
  \item{unit}{supported units, only "mm", "cm", "inches"}

}
\details{
This function coverts units from the first dimension in the data coordinate.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
