\name{cytoband.col}
\alias{cytoband.col}
\title{
Assign colors to cytogenetic band (hg19) according to the Giemsa stain results
}
\description{
Assign colors to cytogenetic band (hg19) according to the Giemsa stain results
}
\usage{
cytoband.col(x)
}
\arguments{

  \item{x}{A vector containing the Giemsa stain results}

}
\examples{
# There is no example
NULL

}
