\name{fontsize}
\alias{fontsize}
\title{
Convert fontsize to cex
}
\description{
Convert fontsize to cex
}
\usage{
fontsize(x)
}
\arguments{

  \item{x}{value for fontsize}

}
\details{
It just mark the value as a fontsize setting for \code{\link{circos.text}}

}
\author{
Zuguang Gu <z.gu@dkfz.de>

}
\examples{
# There is no example
NULL

}
