\name{get.all.sector.index}
\alias{get.all.sector.index}
\title{
  Get index for all sectors  


}
\description{
  Get index for all sectors  


}
\usage{
get.all.sector.index()
}
\details{
  Simple function returning a vector of all sector index. 


}
\references{
Gu, Z. (2014) circlize implements and enhances circular visualization in R. Bioinformatics.
}
