\name{inches_x}
\alias{inches_x}
\title{
Convert unit on x direction in data coordinate
}
\description{
Convert unit on x direction in data coordinate
}
\usage{
inches_x(...)
}
\arguments{

  \item{...}{pass to \code{\link{convert_x}}}

}
\details{
This function is same as \code{\link{convert_x}} in inch unit.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# see example in `convert_x` page
NULL
}
