\name{inches_y}
\alias{inches_y}
\title{
Convert unit on y direction in data coordinate
}
\description{
Convert unit on y direction in data coordinate
}
\usage{
inches_y(...)
}
\arguments{

  \item{...}{pass to \code{\link{convert_y}}}

}
\details{
This function is same as \code{\link{convert_y}} in inch unit.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# see example in `convert_y` page
NULL
}
