\name{mm_h}
\alias{mm_h}
\title{
Convert units
}
\description{
Convert units
}
\usage{
mm_h(...)
}
\arguments{

  \item{...}{pass to \code{\link{convert_length}}}

}
\details{
This function is same as \code{\link{convert_length}} in mm unit.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# see example in `convert_length` page
NULL
}
