\name{show.index}
\alias{show.index}
\title{
  Label the sector index and the track index of each cell


}
\description{
  Label the sector index and the track index of each cell


}
\usage{
show.index()
}
\details{
  Draw the index of the sector and track for each cell on the figure.This function can help you to find the coordinate of cells. 


}
