\name{smartAlign}
\alias{smartAlign}
\title{
Adjust positions of rectanglar shapes
}
\description{
Adjust positions of rectanglar shapes
}
\usage{
smartAlign(bottom, top, xlim)
}
\arguments{

  \item{bottom}{position which corresponds to the bottom of the rectangle-shapes}
  \item{top}{position which corresponds to the top of the rectanglar shapes}
  \item{xlim}{ranges on x-axis}

}
\details{
It adjust the position of the rectangular shapes to make them do not overlap
}
\examples{
# There is no example
NULL

}
