\name{uh}
\alias{uh}
\title{
Convert units
}
\description{
Convert units
}
\usage{
uh(...)
}
\arguments{

  \item{...}{pass to \code{\link{convert_length}}}

}
\details{
This function is same as \code{\link{convert_length}}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# see example in `convert_length` page
NULL
}
