\name{ux}
\alias{ux}
\title{
Convert unit on x direction in data coordinate
}
\description{
Convert unit on x direction in data coordinate
}
\usage{
ux(...)
}
\arguments{

  \item{...}{pass to \code{\link{convert_x}}.}

}
\details{
Please do not use this function. Use \code{\link{mm_x}}/\code{\link{cm_x}}/inches_x` instead.
}
\examples{
# There is no example
NULL

}
