\name{uy}
\alias{uy}
\title{
Convert unit on y direction in data coordinate
}
\description{
Convert unit on y direction in data coordinate
}
\usage{
uy(...)
}
\arguments{

  \item{...}{pass to \code{\link{convert_y}}}

}
\details{
This function is same as \code{\link{convert_y}}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# see example in `convert_y` page
NULL
}
