\name{adjacencyList2Matrix}
\alias{adjacencyList2Matrix}
\title{
Convert adjacency list to adjacency matrix
}
\description{
Convert adjacency list to adjacency matrix
}
\usage{
adjacencyList2Matrix(lt, square = FALSE)
}
\arguments{

  \item{lt}{a data frame which contains adjacency list.}
  \item{square}{is the returned matrix a square matrix?}

}
\details{
Convert adjacency list to adjacency matrix.
}
\examples{
# There is no example
NULL

}
