\name{circos.clear}
\alias{circos.clear}
\title{
  Reset the circos layout parameters


}
\description{
  Reset the circos layout parameters


}
\usage{
circos.clear()
}
\details{
  The package uses several global variables to trace the plotting procedure.If you want to draw a new figure, you must call this function first to reseteverything.


}
