\name{circos.clear}
\alias{circos.clear}
\title{
Reset the circular layout parameters
}
\description{
Reset the circular layout parameters
}
\usage{
circos.clear()
}
\details{
Because there are several
parameters for the circular plot which can only be set before \code{\link{circos.initialize}}. So before you draw the next
circular plot, you need to reset all these parameters.

If you meet some errors when re-drawing the circular plot, try running this function and it will solve most of the problems.
}
\examples{
# There is no example
NULL

}
