\name{circos.clear}
\alias{circos.clear}
\title{
  Reset the circos layout parameters and internal variables


}
\description{
  Reset the circos layout parameters and internal variables


}
\usage{
circos.clear()
}
\details{
  The package uses several global variables (internal) to trace the plotting procedure.If you want to draw a new figure, you must call this function first to reseteverything.


}
