\name{cm_h}
\alias{cm_h}
\title{
Convert units
}
\description{
Convert units
}
\usage{
cm_h(h)
}
\arguments{

  \item{h}{The height in numeric.}

}
\details{
See explanations in \code{\link{convert_length}} page.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# see examples in `convert_length` page
NULL
}
