\name{cm_x}
\alias{cm_x}
\title{
Convert unit on x direction in data coordinate
}
\description{
Convert unit on x direction in data coordinate
}
\usage{
cm_x(x, sector.index = get.current.sector.index(),
    track.index = get.current.track.index(), ...)
}
\arguments{

  \item{x}{The x-value in numeric.}
  \item{sector.index}{Index of sector.}
  \item{track.index}{Index of track.}
  \item{...}{Pass to \code{\link{convert_x}}.}

}
\details{
See explanations in \code{\link{convert_x}} page.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# see examples in `convert_x` page
NULL
}
