\name{cm_y}
\alias{cm_y}
\title{
Convert unit on y direction in data coordinate
}
\description{
Convert unit on y direction in data coordinate
}
\usage{
cm_y(y, sector.index = get.current.sector.index(),
    track.index = get.current.track.index())
}
\arguments{

  \item{y}{The y-value in numeric.}
  \item{sector.index}{Index of sector.}
  \item{track.index}{Index of track.}

}
\details{
See explanations in \code{\link{convert_y}} page.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# see examples in `convert_y` page
NULL
}
