\name{convert_height}
\alias{convert_height}
\title{
Convert units
}
\description{
Convert units
}
\usage{
convert_height(...)
}
\arguments{

  \item{...}{pass to \code{\link{convert_length}}}

}
\details{
This function is same as \code{\link{convert_length}}. The reason for naming this function
is \code{\link{convert_length}} is mostely used for defining the height of tracks and track margins.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# see example in `convert_length` page
NULL
}
