\name{get.current.chromosome}
\alias{get.current.chromosome}
\title{
Get current chromosome name
}
\description{
Get current chromosome name
}
\usage{
get.current.chromosome()
}
\details{
The function is same as \code{\link{get.current.sector.index}} and
should only be put inside \code{panel.fun} when using \code{\link{circos.genomicTrackPlotRegion}}.
}
\examples{
# There is no example
NULL

}
