\name{highlight.chromosome}
\alias{highlight.chromosome}
\title{
Highlight chromosomes
}
\description{
Highlight chromosomes
}
\usage{
highlight.chromosome(...)
}
\arguments{

  \item{...}{pass to \code{\link{highlight.sector}}}

}
\details{
This is only a shortcut function of \code{\link{highlight.sector}}.
}
\examples{
# There is no example
NULL

}
