\name{inch_x}
\alias{inch_x}
\title{
Convert unit on x direction in data coordinate
}
\description{
Convert unit on x direction in data coordinate
}
\usage{
inch_x(...)
}
\arguments{

  \item{...}{pass to \code{\link{inches_x}}.}

}
\details{
This function is the same as \code{\link{inches_x}}.
}
\examples{
# There is no example
NULL

}
