\name{inch_y}
\alias{inch_y}
\title{
Convert unit on y direction in data coordinate
}
\description{
Convert unit on y direction in data coordinate
}
\usage{
inch_y(...)
}
\arguments{

  \item{...}{pass to \code{\link{inches_y}}}

}
\details{
This function is the same as \code{\link{inches_y}}.
}
\examples{
# There is no example
NULL

}
