\name{rand_color}
\alias{rand_color}
\title{
  generate random colors  


}
\description{
  generate random colors  


}
\usage{
rand_color(n = 1, transparency = 0)
}
\arguments{
  \item{n}{number of colors}
  \item{transparency}{transparency, numeric value between 0 and 1}

}
\value{
  a vector of colors 


}
