\name{set.current.cell}
\alias{set.current.cell}
\title{
Set flag to current cell
}
\description{
Set flag to current cell
}
\usage{
set.current.cell(sector.index, track.index)
}
\arguments{

  \item{sector.index}{sector index}
  \item{track.index}{track index}

}
\examples{
# There is no example
NULL
}
