\name{show.index}
\alias{show.index}
\title{
Label the sector index and the track index on each cell
}
\description{
Label the sector index and the track index on each cell
}
\usage{
show.index()
}
\details{
This function is deprecated, please use \code{\link{circos.info}} instead.
}
\examples{
# There is no example
NULL

}
