\name{show.index}
\alias{show.index}
\title{
  Label the sector index and the track index on each cell  


}
\description{
  Label the sector index and the track index on each cell  


}
\usage{
show.index()
}
\details{
  This function is deprecated, please use \code{\link{circos.info}} instead. 


}
\references{
Gu, Z. (2014) circlize implements and enhances circular visualization in R. Bioinformatics.
}
