\name{smartAlign}
\alias{smartAlign}
\title{
Adjust positions of text
}
\description{
Adjust positions of text
}
\usage{
smartAlign(x1, x2, xlim)
}
\arguments{

  \item{x1}{Position which corresponds to the top of the text.}
  \item{x2}{Position which corresponds to the bottom of the text.}
  \item{xlim}{Ranges on x-axis.}

}
\details{
used internally
}
\examples{
# There is no example
NULL

}
