\name{uh}
\alias{uh}
\title{
Convert units
}
\description{
Convert units
}
\usage{
uh(...)
}
\arguments{

  \item{...}{pass to \code{\link{convert_length}}.}

}
\details{
Please do not use this function. Use \code{\link{mm_h}}/\code{\link{cm_h}}/inches_h` instead.
}
\examples{
# There is no example
NULL

}
