\name{uy}
\alias{uy}
\title{
Convert unit on y direction in data coordinate
}
\description{
Convert unit on y direction in data coordinate
}
\usage{
uy(...)
}
\arguments{

  \item{...}{pass to \code{\link{convert_y}}.}

}
\details{
Please do not use this function. Use \code{\link{mm_y}}/\code{\link{cm_y}}/inches_y` instead.
}
\examples{
# There is no example
NULL

}
